Add configure option for how read fails behave
Different platforms have different requirements for handling
hwmon sysfs access failures.

The default behavior is that a read failure leads to the daemon
exiting after some number of read failures.  This can be controlled
to then remove the object from the dbus on failure.  Now, it can
instead be controlled to return -errno on read failure.

This new behavior to set the value to -errno must be checked by
whatever mechanism is reporting the value.

Change-Id: I50b93aea56f22267da79c9571319f281e24a1e6f
Signed-off-by: Patrick Venture <venture@google.com>
Signed-off-by: Doyle Huang <doyle.sy.huang@mail.foxconn.com>