Add findHwmonFromDevPath()
This function can find the hwmon path from a /devices
path.  This path is similar to what udev lists as the
device path, but with the /hwmon/hwmonN directories
removed from the end.  In addition, any ':' characters
in the path have been converted to '--', so the code
needs to convert them back.

This is done because some devices are not in the device
tree, and hence wouldn't have the open firmware device
path that the code previously only supported.

Tested: Tested with the new /devices path.

Change-Id: Ic26646c6f609323cff86d4cd10bbe0a44d7d61ac
Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Signed-off-by: Doyle Huang <doyle.sy.huang@mail.foxconn.com>